Refactor to replace `Person user` by `ObservablePerson user` in the `ResumeBook`
