+++ src/main/java/seedu/address/logic/Logic.java
@@ -9,8 +9,7 @@ import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReadOnlyResumeBook;
 import seedu.address.model.item.Item;
-import seedu.address.model.item.ObservablePerson;
-import seedu.address.model.item.Person;
+import seedu.address.model.item.wrapper.ObservablePerson;
 
 /**
  * API of the Logic component
@@ -50,7 +49,5 @@ public interface Logic {
      */
     void setGuiSettings(GuiSettings guiSettings);
 
-    Person getUser();
-
-    ObservablePerson getObservablePerson();
+    ObservablePerson getUser();
 }
