+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -15,8 +15,7 @@ import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyResumeBook;
 import seedu.address.model.item.Item;
-import seedu.address.model.item.ObservablePerson;
-import seedu.address.model.item.Person;
+import seedu.address.model.item.wrapper.ObservablePerson;
 import seedu.address.storage.Storage;
 
 /**
@@ -64,16 +63,6 @@ public class LogicManager implements Logic {
     }
 
     @Override
-    public Person getUser() {
-        return model.getUser();
-    }
-
-    @Override
-    public ObservablePerson getObservablePerson() {
-        return model.getObservablePerson();
-    }
-
-    @Override
     public Path getAddressBookFilePath() {
         return model.getResumeBookFilePath();
     }
@@ -86,5 +75,10 @@ public class LogicManager implements Logic {
     @Override
     public void setGuiSettings(GuiSettings guiSettings) {
         model.setGuiSettings(guiSettings);
+    }
+
+    @Override
+    public ObservablePerson getUser() {
+        return model.getUser();
     }
 }
