+++ src/main/java/seedu/address/logic/commands/ResumePreviewCommand.java
@@ -79,7 +79,7 @@ public class ResumePreviewCommand extends Command {
         data.append("\n=========================\n")
                 .append("PERSONAL DETAILS\n")
                 .append("=========================\n\n")
-                .append(model.getUser().toPreview())
+                .append(model.getUser().getInternalPerson().toPreview())
                 .append("\n\n");
 
         data.append("=========================\n")
