+++ src/main/java/seedu/address/logic/commands/me/EditUserCommand.java
@@ -71,7 +71,7 @@ public class EditUserCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        Person userToEdit = model.getUser();
+        Person userToEdit = model.getUser().getInternalPerson();
 
         Person editedUser = createEditedUser(userToEdit, editUserDescriptor);
         model.setUser(editedUser);
