+++ src/main/java/seedu/address/model/Model.java
@@ -9,11 +9,11 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.item.Internship;
 import seedu.address.model.item.Item;
-import seedu.address.model.item.ObservablePerson;
 import seedu.address.model.item.Person;
 import seedu.address.model.item.Project;
 import seedu.address.model.item.Resume;
 import seedu.address.model.item.Skill;
+import seedu.address.model.item.wrapper.ObservablePerson;
 
 /**
  * The API of the Model component.
@@ -71,9 +71,9 @@ public interface Model {
     void setUser(Person person);
 
     /**
-     * @return Person in the model.
+     * @return ObservablePerson in the model.
      */
-    Person getUser();
+    ObservablePerson getUser();
 
 
     //=========== Internships ================================================================================
@@ -320,6 +320,4 @@ public interface Model {
      * Saves the current resume book state for undo/redo.
      */
     void commitResumeBook();
-
-    ObservablePerson getObservablePerson();
 }
