+++ src/main/java/seedu/address/model/item/wrapper/UniqueItemList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.item;
+package seedu.address.model.item.wrapper;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -10,6 +10,7 @@ import java.util.stream.Collectors;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import seedu.address.model.item.Item;
 import seedu.address.model.item.exceptions.DuplicateItemException;
 import seedu.address.model.item.exceptions.ItemNotFoundException;
